pyproject.toml
[project]
name = "hr-resource-chatbot"
version = "0.1.0"
description = "HR assistant chatbot with retrieval-augmented generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic>=2.7",
    "sentence-transformers>=3.0.1",
    "faiss-cpu>=1.7.4.post2",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "google-genai>=0.3.0",  # official Google GenAI Python SDK
]
